chatbot/requirements.txt
# Core dependencies
openai>=1.0.0
chromadb>=0.4.0
redis>=5.0.0
python-dotenv>=1.0.0

# Web dependencies
selenium>=4.0.0
requests>=2.0.0
webdriver_manager>=4.0.0

# Data processing
pandas>=2.0.0
numpy>=1.0.0
textblob>=0.0.0

# Vector embeddings
sentence-transformers>=2.0.0
torch>=2.0.0

# Async support
aiohttp>=3.0.0
asyncio>=3.4.3

# Utilities
python-dateutil>=2.0.0
typing-extensions>=4.0.0

# 原来缺失，后补上的
tiktoken
langdetect